use the fragment class to replace the activity class in this newbranch
